refactor: 기존 api 로직 legacy로 이동
