+++ api/legacy/auth/index.ts
