+++ api/legacy/hooks/members.ts
@@ -7,8 +7,8 @@ import {
   getMemberProfileOfMe,
   getMembersSearchByName,
   postMemberCoffeeChat,
-} from '@/api/members';
-import { PostMemberCoffeeChatVariables, ProfileDetail } from '@/api/members/type';
+} from '@/api/legacy/members';
+import { PostMemberCoffeeChatVariables, ProfileDetail } from '@/api/legacy/members/type';
 
 // 멤버 프로필 조회
 export const useGetMemberOfMe = () => {
