+++ api/legacy/makers/index.ts
@@ -1,5 +1,5 @@
 import { axiosInstance } from '@/api';
-import { Maker } from '@/api/makers/types';
+import { Maker } from '@/api/legacy/makers/types';
 
 export const getMakersProfile = async () => {
   const { data } = await axiosInstance.request<Maker[]>({
