+++ components/projects/upload/constants.ts
@@ -1,6 +1,6 @@
 import { DefaultValues } from 'react-hook-form';
 
-import { LinkTitle } from '@/api/projects/type';
+import { LinkTitle } from '@/api/legacy/projects/type';
 import { DEFAULT_MEMBER } from '@/components/projects/upload/MemberForm/constants';
 import { Category, FormItem } from '@/components/projects/upload/types';
 import { ProjectUploadForm } from '@/pages/projects/upload';
