+++ pages/makers/index.tsx
@@ -3,7 +3,7 @@ import { GetStaticProps } from 'next';
 import { FC } from 'react';
 import { RemoveScroll } from 'react-remove-scroll';
 
-import { getMakersProfile } from '@/api/makers';
+import { getMakersProfile } from '@/api/legacy/makers';
 import Footer from '@/components/common/Footer';
 import SwitchableHeader from '@/components/common/Header/SwitchableHeader';
 import AboutMakers from '@/components/makers/AboutMakers';
